requirements.txt
openai~=1.75.0
python-dotenv
djangorestframework
requests
django
django-cors-headers
channels
uvicorn